frontend/requirements.txt
streamlit>=1.36
requests>=2.31
boto3>=1.34
pandas>=2.0
python-dotenv>=1.0
